[IMP] survey: improve label axis x in chartjs live survey
Before when there were long labels, this was leading to an overlapping
of the text.

Now, we display the whole label but in multiline with a maximum number of
characters per line depending on the number of columns.
The font size will also be compute depending of the number of characters
per line and with the length of the longest word in the labels.

task-2382681

closes odoo/odoo#64877

X-original-commit: 58bd6a51041d1c0c011810823caf540995428c2a
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>