[FIX] web_editor: fix scroll event on snippet popup
Before this commit, when scrolling a snippet popup, the overlay was
not updating its position.

So now, we use addEventListener instead of jQuery for the scroll event
on the scrolling element because we need the 'capture' parameter.
Setting this capture parameter to true allows us to take advantage of
event bubbling for events that otherwise don’t support it.

Part of https://github.com/odoo/odoo/pull/61998
task-2312878

X-original-commit: e734df614a07d51cfd2c9eb6e3876189e1d22a69
Co-authored-by: Samuel Degueldre <sad@odoo.com>